add modules with functions and variables
