+++ src/modules/Variables.js
@@ -1,0 +1,4 @@
+const message = document.querySelector('.message');
+const refreshButton = document.querySelector('#refresh');
+const form = document.querySelector('form');
+export { message, refreshButton, form };
